Move fixture classes to their own files
This makes the code compliant with the new "psr4" cs-fixer rule.
