+++ tests/Admin/BaseAdminModelManager_Admin.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\AdminBundle\Tests\Admin;
+
+use Sonata\AdminBundle\Admin\AbstractAdmin;
+
+class BaseAdminModelManager_Admin extends AbstractAdmin
+{
+}
