+++ tests/Controller/AdminControllerHelper_Bar.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\AdminBundle\Tests\Controller;
+
+class AdminControllerHelper_Bar
+{
+    public function getAdminTitle()
+    {
+        return 'bar';
+    }
+
+    public function setEnabled($value)
+    {
+    }
+
+    public function getEnabled()
+    {
+    }
+}
